requirements.txt
Flask
google-generativeai
pillow